+++ app/src/main/java/in/hridayan/ashell/adapters/AboutAdapter.java
@@ -167,6 +167,11 @@ public class AboutAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> 
               ((FragmentActivity) activity)
                   .getSupportFragmentManager()
                   .beginTransaction()
+                  .setCustomAnimations(
+                      R.anim.fragment_enter,
+                      R.anim.fragment_exit,
+                      R.anim.fragment_pop_enter,
+                      R.anim.fragment_pop_exit)
                   .replace(R.id.fragment_container, new ChangelogFragment())
                   .addToBackStack(null)
                   .commit();
