+++ app/src/main/java/in/hridayan/ashell/adapters/ExamplesAdapter.java
@@ -8,7 +8,6 @@ import static in.hridayan.ashell.utils.Preferences.SORT_Z_TO_A;
 import android.animation.ObjectAnimator;
 import android.app.Activity;
 import android.content.Context;
-import android.content.Intent;
 import android.content.res.Configuration;
 import android.os.Build;
 import android.view.LayoutInflater;
@@ -16,21 +15,17 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.view.animation.AnimationUtils;
 import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.AppCompatImageButton;
-import androidx.fragment.app.FragmentActivity;
 import androidx.recyclerview.widget.RecyclerView;
 import com.google.android.material.card.MaterialCardView;
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import com.google.android.material.textview.MaterialTextView;
 import in.hridayan.ashell.R;
-import in.hridayan.ashell.UI.MainViewModel;
-import in.hridayan.ashell.activities.MainActivity;
-import in.hridayan.ashell.fragments.aShellFragment;
 import in.hridayan.ashell.utils.CommandItems;
 import in.hridayan.ashell.utils.HapticUtils;
 import in.hridayan.ashell.utils.Preferences;
 import in.hridayan.ashell.utils.Utils;
+import in.hridayan.ashell.viewmodels.MainViewModel;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
