+++ app/src/main/java/in/hridayan/ashell/fragments/ExamplesFragment.java
@@ -33,8 +33,6 @@ import com.google.android.material.search.SearchBar;
 import com.google.android.material.textview.MaterialTextView;
 import in.hridayan.ashell.R;
 import in.hridayan.ashell.UI.CustomSearchView;
-import in.hridayan.ashell.UI.ExamplesViewModel;
-import in.hridayan.ashell.UI.MainViewModel;
 import in.hridayan.ashell.adapters.CommandsSearchAdapter;
 import in.hridayan.ashell.adapters.ExamplesAdapter;
 import in.hridayan.ashell.utils.CommandItems;
@@ -43,6 +41,8 @@ import in.hridayan.ashell.utils.HapticUtils;
 import in.hridayan.ashell.utils.Preferences;
 import in.hridayan.ashell.utils.ThemeUtils;
 import in.hridayan.ashell.utils.Utils;
+import in.hridayan.ashell.viewmodels.ExamplesViewModel;
+import in.hridayan.ashell.viewmodels.MainViewModel;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
