+++ app/src/main/java/in/hridayan/ashell/fragments/otgShellFragment.java
@@ -66,7 +66,6 @@ import in.hridayan.ashell.UI.BehaviorFAB.FabOtgScrollUpListener;
 import in.hridayan.ashell.UI.BehaviorFAB.OtgShareButtonListener;
 import in.hridayan.ashell.UI.CoordinatedNestedScrollView;
 import in.hridayan.ashell.UI.KeyboardUtils;
-import in.hridayan.ashell.UI.MainViewModel;
 import in.hridayan.ashell.activities.MainActivity;
 import in.hridayan.ashell.adapters.CommandsAdapter;
 import in.hridayan.ashell.adapters.SettingsAdapter;
@@ -78,6 +77,7 @@ import in.hridayan.ashell.utils.OtgUtils.MessageOtg;
 import in.hridayan.ashell.utils.Preferences;
 import in.hridayan.ashell.utils.SettingsItem;
 import in.hridayan.ashell.utils.Utils;
+import in.hridayan.ashell.viewmodels.MainViewModel;
 import java.io.File;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
