+++ app/src/main/java/in/hridayan/ashell/viewmodels/AboutViewModel.java
@@ -1,4 +1,4 @@
-package in.hridayan.ashell.UI;
+package in.hridayan.ashell.viewmodels;
 
 import androidx.lifecycle.ViewModel;
