+++ app/src/main/res/anim/fragment_pop_exit.xml
@@ -12,7 +12,7 @@
         android:startOffset="66"
         android:toAlpha="0.0" />
     <scale
-        android:duration="300"
+        android:duration="350"
         android:fillAfter="true"
         android:fillBefore="true"
         android:fillEnabled="true"
